Add cluster_position.py, a class for computing cluster positions and shapes.
Erik's code does not include an estimate for the mean position of the cluster, so this class is meant to be used for that.

The contents of this class were mainly pulled from the dbscan_analysis.ComputeClusterPosition method in dbscan_analysis.py. I will
delete that method and instead add an interface to this one.